moved getUserDataFromDocument etc. into server-tools.js
