+++ src/editor-state-prototypes.js
@@ -33,27 +33,8 @@ function getUrlParams() {
   const report = queryParams.viewMode === 'report'
   return { viewMode: report ? 'REPORT_VIEW' : 'EXPLORE_VIEW' }
 }
-
-function getUserDataFromDocument() {
-  const userData = document.getElementById('userData')
-  if (userData) {
-    return { userData: JSON.parse(userData.textContent) }
-  }
-  return {}
-}
-
-function getNotebookInfoFromDocument() {
-  const notebookInfo = document.getElementById('notebookInfo')
-  if (notebookInfo) {
-    return { notebookInfo: JSON.parse(notebookInfo.textContent) }
-  }
-  return {}
-}
-
 export {
   getUrlParams,
-  getUserDataFromDocument,
-  getNotebookInfoFromDocument,
   newCell,
   newCellID,
   newNotebook,
